# backend/openapi.yaml
openapi: 3.0.3
info:
  title: WEBTE2 API
  version: 1.0.0
  description: REST API for Octave sessions, simulations, logs, statistics, and API documentation export.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    StatsTrackRequest:
      type: object
      properties:
        animationType:
          type: string
          enum:
            - pendulum
            - ballbeam
      required:
        - animationType
    StatsTrackResponse:
      type: object
      properties:
        tracked:
          type: boolean
        reason:
          type: string
          nullable: true
      required:
        - tracked
    StatsSummary:
      type: object
      properties:
        pendulum:
          type: integer
        ballbeam:
          type: integer
      required:
        - pendulum
        - ballbeam
    StatsDetailItem:
      type: object
      properties:
        animation_type:
          type: string
          enum:
            - pendulum
            - ballbeam
        user_token:
          type: string
          format: uuid
        city:
          type: string
        country:
          type: string
        used_at:
          type: string
          format: date-time
      required:
        - animation_type
        - user_token
        - city
        - country
        - used_at
security:
  - bearerAuth: []
paths:
...
  /api/stats/track:
    post:
      summary: Track animation usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatsTrackRequest'
      responses:
        '200':
          description: Tracking result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsTrackResponse'
        '400':
          description: Invalid animation type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Tracking failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stats:
    get:
      summary: Get animation usage totals
      responses:
        '200':
          description: Usage counters by animation type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stats/details:
    get:
      summary: Get detailed animation usage entries
      responses:
        '200':
          description: Detailed usage list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatsDetailItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load stats details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...